+++ c3IDE/Windows/ConditionWindow.xaml.cs
@@ -41,6 +41,7 @@ namespace c3IDE.Windows
         private CompletionWindow completionWindow;
         private FoldingManager aceFoldingManager;
         private BraceFoldingStrategy folding;
+        private SearchPanel acePanel, langPanel, codePanel;
 
         /// <summary>
         /// condition window constructor 
@@ -68,9 +69,9 @@ namespace c3IDE.Windows
             folding.UpdateFoldings(aceFoldingManager, CodeTextEditor.Document);
 
             //setip ctrl-f to single page code find
-            SearchPanel.Install(CodeTextEditor);
-            SearchPanel.Install(LanguageTextEditor);
-            SearchPanel.Install(AceTextEditor);
+            codePanel = SearchPanel.Install(CodeTextEditor);
+            langPanel = SearchPanel.Install(LanguageTextEditor);
+            acePanel = SearchPanel.Install(AceTextEditor);
         }
 
         /// <summary>
@@ -81,6 +82,7 @@ namespace c3IDE.Windows
             ThemeManager.SetupTextEditor(AceTextEditor, Syntax.Json);
             ThemeManager.SetupTextEditor(LanguageTextEditor, Syntax.Json);
             ThemeManager.SetupTextEditor(CodeTextEditor, Syntax.Javascript);
+            ThemeManager.SetupSearchPanel(acePanel, langPanel, codePanel);
 
             if (AddonManager.CurrentAddon != null)
             {